+++ elmyra.ip.access.epo/CHANGES.rst
@@ -7,6 +7,7 @@ development
 - middleware: propagate userid upstream to middleware and resolve user detail information from MongoDB
 - middleware: use OPS credentials from user details, otherwise fall back to Elmyra OPS credentials
 - middleware: publish and enhance OPS usage api
+- middleware: per-user, per-day metrics for upstream transfer volume
 
 0.71.4
 ======